feat(poc): add chain-side validations and nonce type change for Phase 3.5
Changes:
- Add strict count increase validation for PoCV2StoreCommit (reject if count <= last)
- Add same-block rate limit (only one commit per block allowed)
- Add exact sum validation for MLNodeWeightDistribution (sum must equal committed count)
- Add commit_block_height field to PoCV2StoreCommit storage type
- Change PoCArtifactV2.nonce from int64 to int32
- Add getWeightDistribution() to proportionally adjust weights when count mismatch
- Update documentation for Phase 3.5 completion

+++ decentralized-api/internal/pocv2/node_orchestrator.go
@@ -93,7 +93,7 @@ func (b *OrchestratorChainBridgeV2Impl) PoCv2BatchesForStage(startPoCBlockHeight
 			artifacts := make([]*ArtifactV2, 0, len(chainBatch.Artifacts))
 			for _, artifact := range chainBatch.Artifacts {
 				artifacts = append(artifacts, &ArtifactV2{
-					Nonce:     artifact.Nonce,
+					Nonce:     int64(artifact.Nonce),
 					VectorB64: base64.StdEncoding.EncodeToString(artifact.Vector), // Chain stores as bytes, convert to base64 string
 				})
 			}
